#227 moved shared files to comon
