+++ packages/common/package.json
@@ -1,0 +1,51 @@
+{
+  "private": true,
+  "scripts": {
+    "test": "istanbul cover _mocha test -- --opts test/mocha.opts"
+  },
+  "dependencies": {
+    "assign-styles": "2.0.0",
+    "inline-style-transformer": "^1.0.0",
+    "lodash": "^4.2.1"
+  },
+  "peerDependencies": {
+    "react": "^0.14.3",
+    "react-dom": "^0.14.3"
+  },
+  "devDependencies": {
+    "babel": "^6.3.26",
+    "babel-cli": "^6.4.0",
+    "babel-core": "^6.4.0",
+    "babel-loader": "^6.2.4",
+    "babel-plugin-add-module-exports": "^0.1.2",
+    "babel-plugin-lodash": "^2.1.0",
+    "babel-preset-es2015": "^6.3.13",
+    "babel-preset-react": "^6.3.13",
+    "babel-preset-stage-0": "^6.5.0",
+    "chai": "^3.4.1",
+    "eslint": "^1.10.3",
+    "eslint-config-rackt": "^1.1.1",
+    "eslint-plugin-react": "^3.15.0",
+    "istanbul": "^1.0.0-alpha.2",
+    "jsdom": "^7.2.2",
+    "lodash": "^4.3.0",
+    "mocha": "^2.3.4",
+    "react": "^0.14.6",
+    "react-addons-test-utils": "^0.14.7",
+    "react-dom": "^0.14.6",
+    "rimraf": "^2.5.0",
+    "sinon": "^1.17.2",
+    "sinon-chai": "^2.8.0"
+  },
+  "babel": {
+    "presets": [
+      "react",
+      "es2015",
+      "stage-0"
+    ],
+    "plugins": [
+      "lodash",
+      "add-module-exports"
+    ]
+  }
+}
