+++ packages/common/plugins/styleLogger.js
