+++ packages/common/test/utils/splitCondition-test.js
@@ -1,27 +1,26 @@
-import splitCondition from '../../modules/utils/splitCondition'
-import { expect } from 'chai'
+import splitCondition from '../../utils/splitCondition'
 
 describe('Splitting conditions', () => {
 
   it('should replace left by props value and return left and right', () => {
-    const Component = {props: {active: true}}
-    const output = {left: 'true', right: 'true'}
+    const Component = { props: { active: true } }
+    const output = { left: 'true', right: 'true' }
 
     expect(splitCondition('active=true', '=', Component)).to.eql(output)
   })
 
   it('should resolve number values as numbers', () => {
-    const Component = {props: {clicks: 3}}
-    const output = {left: 3, right: '2'}
+    const Component = { props: { clicks: 3 } }
+    const output = { left: 3, right: '2' }
 
     expect(splitCondition('clicks>=2', '>=', Component)).to.eql(output)
   })
 
   it('should return false if left is not part of props or state', () => {
-    expect(splitCondition('active=true', '=', {})).to.eql(false)
+    expect(splitCondition('active=true', '=', { })).to.eql(false)
   })
 
   it('should return false if the operator does not match', () => {
-    expect(splitCondition('active=true', '!=', {})).to.eql(false)
+    expect(splitCondition('active=true', '!=', { })).to.eql(false)
   })
 })
