+++ packages/common/utils/copyProperties.js
