+++ packages/common/utils/getChildType.js
