+++ packages/common/utils/getPseudoExpression.js
