+++ packages/common/utils/splitCondition.js
